requirements.txt
crewai
crewai-tools
dash
requests
pydub
python-dotenv
openai
openai-whisper
transformers
torchaudio
dash_table
dash
